Get `pnpm dev` working!
