+++ packages/math-input/less/main.less
@@ -1,2 +1,2 @@
 @import "./overrides.less";
-@import (inline) "../../../node_modules/mathquill/build/mathquill.css";
+@import (inline) "../node_modules/mathquill/build/mathquill.css";
